Bump firmware version
