+++ Meshtastic/Views/Helpers/Weather/IAQScale.swift
@@ -1,0 +1,18 @@
+//
+//  IAQScale.swift
+//  Meshtastic
+//
+//  Created by Garth Vander Houwen on 4/24/24.
+//
+
+import SwiftUI
+
+struct IAQScale: View {
+
+	let gradient = Gradient(colors: [.green, .mint, .yellow, .orange, .red, .purple, .purple, .brown, .brown, .brown, .brown])
+	var body: some View {
+		ZStack (alignment: .leading) {
+
+		}
+	}
+}
